* Split the initialization, refinement, and extraction of a VisionMarker into separate functions * Converted s16 to f32 quads earlier in the fiducial marker extraction * Fix for when RefineQuadrilateral() fails, and some more error handling * All 96 unit tests pass

+++ coretech/vision/include/anki/vision/robot/fiducialDetection.h
@@ -103,7 +103,8 @@ namespace Anki
     // changes its position by more than maxCornerChange, the original quad and
     // homography are returned.
     //
-    Result RefineQuadrilateral(const Quadrilateral<s16>& initialQuad,
+    Result RefineQuadrilateral(
+      const Quadrilateral<f32>& initialQuad,
       const Array<f32>& initialHomography,
       const Array<u8> &image,
       const f32 squareWidthFraction,
